Migrate widget menu
